refactor: move core PR review logic into GitCogniAgent. Simplify gitcogni_flow
